+++ CHANGELOG.md
@@ -5,11 +5,12 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 
 ## [1.9.13] - Unreleased
 ### Added
+- GHORG_CLONE_SNIPPETS as a way to clone all snippets
 ### Changed
 ### Deprecated
 ### Removed
 ### Fixed
-- refs to deprecated io/ioutil (#415); thanks @testwill
+- Refs to deprecated io/ioutil (#415); thanks @testwill
 ### Security
 
 ## [1.9.12] - 5/15/24