+++ CHANGELOG.md
@@ -9,7 +9,8 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 ### Deprecated
 ### Removed
 ### Fixed
-- refs to deprecated io/ioutil (#415); thanks @testwill
+- Refs to deprecated io/ioutil (#415); thanks @testwill
+- Bitbucket clones without app password; thanks @SparklingAperioso
 ### Security
 
 ## [1.9.12] - 5/15/24